Initial commit
This commit includes wire format parsing, protocol code generation,
resource management, and crates for the core Wayland protocol as
well as the xdg_shell protocol. This commit serves primarily as a
"save point" as I experiment with different strategies of where the
the interface of an object id included in a message becomes known
to the program.